setup security config and update mvn dependencies

+++ score-server/src/main/java/bio/overture/score/server/security/ExpiringOauth2Authentication.java
@@ -7,15 +7,15 @@ import org.springframework.security.oauth2.provider.OAuth2Request;
 
 @Getter
 public class ExpiringOauth2Authentication extends OAuth2Authentication {
-  private final int expiry; // expiry time of the authentication token in seconds
+  private final long expiry; // expiry time of the authentication token in seconds
 
   public ExpiringOauth2Authentication(OAuth2Request storedRequest,
-    Authentication userAuthentication, int expiry) {
+    Authentication userAuthentication, long expiry) {
     super(storedRequest, userAuthentication);
     this.expiry = expiry;
   }
 
-  public static ExpiringOauth2Authentication from(OAuth2Authentication existing, int expiry) {
+  public static ExpiringOauth2Authentication from(OAuth2Authentication existing, long expiry) {
     return new ExpiringOauth2Authentication(existing.getOAuth2Request(), existing.getUserAuthentication(), expiry);
   }
 }
